Crate google-bigquery2

This documentation was generated from bigquery crate version 0.0.1. The original source code can be found on github.

Features

Handle the following Resources with ease from the central hub ...

datasets

delete, get, insert, list, patch and update

jobs

get, get query results, insert, list and query

projects

list

tabledata

insert all and list

tables

delete, get, insert, list, patch and update

Upload supported by ...

insert jobs

Everything else about the bigquery v2 API can be found at the official documentation site.

Not what you are looking for ? Find all other google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.tables().update(...).doit()
let r = hub.tables().insert(...).doit()
let r = hub.tables().list(...).doit()
let r = hub.tables().delete(...).doit()
let r = hub.tables().get(...).doit()
let r = hub.tables().patch(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-bigquery2 = "0.0.1"

A complete example

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate serde;
extern crate "google-bigquery2" as bigquery2;
use bigquery2::Table;
use bigquery2::Result;
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use bigquery2::Bigquery;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Bigquery::new(hyper::Client::new(), auth);
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req: Table = Default::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.tables().update(&req, "projectId", "datasetId", "tableId")
             .doit();
 
match result {
    Result::HttpError(err) => println!("HTTPERROR: {:?}", err),
    Result::MissingAPIKey => println!("Missing API Key"),
    Result::MissingToken => println!("Missing Token"),
    Result::Failure(_) => println!("General Failure (Response doesn't print)"),
    Result::FieldClash(clashed_field) => println!("FIELD CLASH: {:?}", clashed_field),
    Result::JsonDecodeError(err) => println!("Json failed to decode: {:?}", err),
    Result::Success(_) => println!("Success (value doesn't print)"),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

About Uploads and Downlods

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

About Customization/Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

About Parts

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifyable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

About Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are borrowed

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Structs

Bigquery
Central instance to access all Bigquery related resource activities

Dataset
There is no detailed description.

DatasetAccess
[Optional] An array of objects that define dataset access for one or more entities. You can set this property when inserting or updating a dataset in order to control who is allowed to access the data. If unspecified at dataset creation time, BigQuery adds default dataset access for the following entities: access.specialGroup: projectReaders; access.role: READER; access.specialGroup: projectWriters; access.role: WRITER; access.specialGroup: projectOwners; access.role: OWNER; access.userByEmail: [dataset creator email]; access.role: OWNER;

DatasetDeleteCallBuilder
Deletes the dataset specified by the datasetId value. Before you can delete a dataset, you must delete all its tables, either manually or by specifying deleteContents. Immediately after deletion, you can create another dataset with the same name.

DatasetGetCallBuilder
Returns the dataset specified by datasetID.

DatasetInsertCallBuilder
Creates a new empty dataset.

DatasetList
There is no detailed description.

DatasetListCallBuilder
Lists all the datasets in the specified project to which the caller has read access; however, a project owner can list (but not necessarily get) all datasets in his project.

DatasetListDatasets
An array of the dataset resources in the project. Each resource contains basic information. For full information about a particular dataset resource, use the Datasets: get method. This property is omitted when there are no datasets in the project.

DatasetMethodsBuilder
A builder providing access to all methods supported on dataset resources. It is not used directly, but through the Bigquery hub.

DatasetPatchCallBuilder
Updates information in an existing dataset. The update method replaces the entire dataset resource, whereas the patch method only replaces fields that are provided in the submitted dataset resource. This method supports patch semantics.

DatasetReference
There is no detailed description.

DatasetUpdateCallBuilder
Updates information in an existing dataset. The update method replaces the entire dataset resource, whereas the patch method only replaces fields that are provided in the submitted dataset resource.

DefaultDelegate
A delegate with a conservative default implementation, which is used if no other delegate is set.

ErrorProto
There is no detailed description.

GetQueryResultsResponse
There is no detailed description.

Job
There is no detailed description.

JobConfiguration
There is no detailed description.

JobConfigurationExtract
There is no detailed description.

JobConfigurationLink
There is no detailed description.

JobConfigurationLoad
There is no detailed description.

JobConfigurationQuery
There is no detailed description.

JobConfigurationTableCopy
There is no detailed description.

JobGetCallBuilder
Retrieves the specified job by ID.

JobGetQueryResultCallBuilder
Retrieves the results of a query job.

JobInsertCallBuilder
Starts a new asynchronous job.

JobList
There is no detailed description.

JobListCallBuilder
Lists all the Jobs in the specified project that were started by the user. The job list returns in reverse chronological order of when the jobs were created, starting with the most recent job created.

JobListJobs
List of jobs that were requested.

JobMethodsBuilder
A builder providing access to all methods supported on job resources. It is not used directly, but through the Bigquery hub.

JobQueryCallBuilder
Runs a BigQuery SQL query synchronously and returns query results if the query completes within a specified timeout.

JobReference
There is no detailed description.

JobStatistics
There is no detailed description.

JobStatistics2
There is no detailed description.

JobStatistics3
There is no detailed description.

JobStatistics4
There is no detailed description.

JobStatus
There is no detailed description.

JsonObject
Represents a single JSON object.

JsonValue
There is no detailed description.

MethodInfo
Contains information about an API request.

MultiPartReader
Provides a Read interface that converts multiple parts into the protocol identified by RFC2387. Note: This implementation is just as rich as it needs to be to perform uploads to google APIs, and might not be a fully-featured implementation.

ProjectList
There is no detailed description.

ProjectListCallBuilder
Lists the projects to which you have at least read access.

ProjectListProjects
Projects to which you have at least READ access.

ProjectMethodsBuilder
A builder providing access to all methods supported on project resources. It is not used directly, but through the Bigquery hub.

ProjectReference
There is no detailed description.

QueryRequest
There is no detailed description.

QueryResponse
There is no detailed description.

Table
There is no detailed description.

TableCell
Represents a single cell in the result set. Users of the java client can detect whether their value result is null by calling 'com.google.api.client.util.Data.isNull(cell.getV())'.

TableDataInsertAllRequest
There is no detailed description.

TableDataInsertAllRequestRows
The rows to insert.

TableDataInsertAllResponse
There is no detailed description.

TableDataInsertAllResponseInsertErrors
An array of errors for rows that were not inserted.

TableDataList
There is no detailed description.

TableDeleteCallBuilder
Deletes the table specified by tableId from the dataset. If the table contains data, all the data will be deleted.

TableFieldSchema
There is no detailed description.

TableGetCallBuilder
Gets the specified table resource by table ID. This method does not return the data in the table, it only returns the table resource, which describes the structure of this table.

TableInsertCallBuilder
Creates a new, empty table in the dataset.

TableList
There is no detailed description.

TableListCallBuilder
Lists all tables in the specified dataset.

TableListTables
Tables in the requested dataset.

TableMethodsBuilder
A builder providing access to all methods supported on table resources. It is not used directly, but through the Bigquery hub.

TablePatchCallBuilder
Updates information in an existing table. The update method replaces the entire table resource, whereas the patch method only replaces fields that are provided in the submitted table resource. This method supports patch semantics.

TableReference
There is no detailed description.

TableRow
Represents a single row in the result set, consisting of one or more fields.

TableSchema
There is no detailed description.

TableUpdateCallBuilder
Updates information in an existing table. The update method replaces the entire table resource, whereas the patch method only replaces fields that are provided in the submitted table resource.

TabledataInsertAllCallBuilder
Streams data into BigQuery one record at a time without needing to run a load job.

TabledataListCallBuilder
Retrieves table data from a specified set of rows.

TabledataMethodsBuilder
A builder providing access to all methods supported on tabledata resources. It is not used directly, but through the Bigquery hub.

ViewDefinition
There is no detailed description.

Enums

Result
A universal result type used as return for all action method results.

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.

Traits

CallBuilder
Identifies types which represent builders for a particular resource method

Delegate
A trait specifying functionality to help controlling any request performed by the API. The trait has a conservative default implementation.

Hub
Identifies the Hub. There is only one per library, this trait is supposed to make intended use more explicit. The hub allows to access all resource methods more easily.

NestedType
Identifies types which are only used by other types internally. They have no special meaning, this trait just marks them for completeness.

Part
Identifies types which are only used as part of other types, which usually are carrying the Resource trait.

ReadSeek
A utility to specify reader types which provide seeking capabilities too

RequestValue
Identifies types which are used in API requests.

ResponseResult
Identifies types which are used in API responses.

UnusedType
Identifies types which are not actually used by the API This might be a bug within the google API schema.