Module google_bigquery2::api

Structs

AggregateClassificationMetrics
Aggregate metrics for classification/classifier models. For multi-class models, the metrics are either macro-averaged or micro-averaged. When macro-averaged, the metrics are calculated for each label and then an unweighted average is taken of those values. When micro-averaged, the metric is calculated globally by counting the total number of correctly predicted rows.

Argument
Input/output argument of a function or a stored procedure.

ArimaCoefficients
Arima coefficients.

ArimaFittingMetrics
ARIMA model fitting metrics.

ArimaForecastingMetrics
Model evaluation metrics for ARIMA forecasting models.

ArimaModelInfo
Arima model information.

ArimaOrder
Arima order, can be used for both non-seasonal and seasonal parts.

ArimaResult
(Auto-)arima fitting result. Wrap everything in ArimaResult for easier refactoring if we want to use model-specific iteration results.

ArimaSingleModelForecastingMetrics
Model evaluation metrics for a single ARIMA forecasting model.

AuditConfig
Specifies the audit configuration for a service. The configuration determines which permission types are logged, and what identities, if any, are exempted from logging. An AuditConfig must have one or more AuditLogConfigs. If there are AuditConfigs for both allServices and a specific service, the union of the two AuditConfigs is used for that service: the log_types specified in each AuditConfig are enabled, and the exempted_members in each AuditLogConfig are exempted. Example Policy with multiple AuditConfigs: { "audit_configs": [ { "service": "allServices", "audit_log_configs": [ { "log_type": "DATA_READ", "exempted_members": [ "user:jose@example.com" ] }, { "log_type": "DATA_WRITE" }, { "log_type": "ADMIN_READ" } ] }, { "service": "sampleservice.googleapis.com", "audit_log_configs": [ { "log_type": "DATA_READ" }, { "log_type": "DATA_WRITE", "exempted_members": [ "user:aliya@example.com" ] } ] } ] } For sampleservice, this policy enables DATA_READ, DATA_WRITE and ADMIN_READ logging. It also exempts jose@example.com from DATA_READ logging, and aliya@example.com from DATA_WRITE logging.

AuditLogConfig
Provides the configuration for logging a type of permissions. Example: { "audit_log_configs": [ { "log_type": "DATA_READ", "exempted_members": [ "user:jose@example.com" ] }, { "log_type": "DATA_WRITE" } ] } This enables 'DATA_READ' and 'DATA_WRITE' logging, while exempting jose@example.com from DATA_READ logging.

BigQueryModelTraining
There is no detailed description.

Bigquery
Central instance to access all Bigquery related resource activities

BigtableColumn
There is no detailed description.

BigtableColumnFamily
There is no detailed description.

BigtableOptions
There is no detailed description.

BinaryClassificationMetrics
Evaluation metrics for binary classification/classifier models.

BinaryConfusionMatrix
Confusion matrix for binary classification models.

Binding
Associates members with a role.

BqmlIterationResult
There is no detailed description.

BqmlTrainingRun
There is no detailed description.

BqmlTrainingRunTrainingOptions
[Output-only, Beta] Training options used by this training run. These options are mutable for subsequent training runs. Default values are explicitly stored for options not specified in the input query of the first training run. For subsequent training runs, any option not explicitly specified in the input query will be copied from the previous training run.

CategoricalValue
Representative value of a categorical feature.

CategoryCount
Represents the count of a single category within the cluster.

Cluster
Message containing the information about one cluster.

ClusterInfo
Information about a single cluster for clustering model.

Clustering
There is no detailed description.

ClusteringMetrics
Evaluation metrics for clustering models.

ConfusionMatrix
Confusion matrix for multi-class classification models.

ConnectionProperty
There is no detailed description.

CsvOptions
There is no detailed description.

DataSplitResult
Data split result. This contains references to the training and evaluation data tables that were used to train the model.

Dataset
There is no detailed description.

DatasetAccess
[Optional] An array of objects that define dataset access for one or more entities. You can set this property when inserting or updating a dataset in order to control who is allowed to access the data. If unspecified at dataset creation time, BigQuery adds default dataset access for the following entities: access.specialGroup: projectReaders; access.role: READER; access.specialGroup: projectWriters; access.role: WRITER; access.specialGroup: projectOwners; access.role: OWNER; access.userByEmail: [dataset creator email]; access.role: OWNER;

DatasetAccessEntry
There is no detailed description.

DatasetAccessEntryTargetTypes
There is no detailed description.

DatasetDeleteCall
Deletes the dataset specified by the datasetId value. Before you can delete a dataset, you must delete all its tables, either manually or by specifying deleteContents. Immediately after deletion, you can create another dataset with the same name.

DatasetGetCall
Returns the dataset specified by datasetID.

DatasetInsertCall
Creates a new empty dataset.

DatasetList
There is no detailed description.

DatasetListCall
Lists all datasets in the specified project to which you have been granted the READER dataset role.

DatasetListDatasets
An array of the dataset resources in the project. Each resource contains basic information. For full information about a particular dataset resource, use the Datasets: get method. This property is omitted when there are no datasets in the project.

DatasetMethods
A builder providing access to all methods supported on dataset resources. It is not used directly, but through the Bigquery hub.

DatasetPatchCall
Updates information in an existing dataset. The update method replaces the entire dataset resource, whereas the patch method only replaces fields that are provided in the submitted dataset resource. This method supports patch semantics.

DatasetReference
There is no detailed description.

DatasetUpdateCall
Updates information in an existing dataset. The update method replaces the entire dataset resource, whereas the patch method only replaces fields that are provided in the submitted dataset resource.

DestinationTableProperties
There is no detailed description.

DimensionalityReductionMetrics
Model evaluation metrics for dimensionality reduction models.

EncryptionConfiguration
There is no detailed description.

Entry
A single entry in the confusion matrix.

ErrorProto
There is no detailed description.

EvaluationMetrics
Evaluation metrics of a model. These are either computed on all training data or just the eval data based on whether eval data was used during training. These are not present for imported models.

ExplainQueryStage
There is no detailed description.

ExplainQueryStep
There is no detailed description.

Explanation
Explanation for a single feature.

Expr
Represents a textual expression in the Common Expression Language (CEL) syntax. CEL is a C-like expression language. The syntax and semantics of CEL are documented at https://github.com/google/cel-spec. Example (Comparison): title: "Summary size limit" description: "Determines if a summary is less than 100 chars" expression: "document.summary.size() < 100" Example (Equality): title: "Requestor is owner" description: "Determines if requestor is the document owner" expression: "document.owner == request.auth.claims.email" Example (Logic): title: "Public documents" description: "Determine whether the document should be publicly visible" expression: "document.type != 'private' && document.type != 'internal'" Example (Data Manipulation): title: "Notification string" description: "Create a notification string with a timestamp." expression: "'New message received at ' + string(document.create_time)" The exact variables and functions that may be referenced within an expression are determined by the service that evaluates it. See the service documentation for additional information.

ExternalDataConfiguration
There is no detailed description.

FeatureValue
Representative value of a single feature within the cluster.

GetIamPolicyRequest
Request message for GetIamPolicy method.

GetPolicyOptions
Encapsulates settings provided to GetIamPolicy.

GetQueryResultsResponse
There is no detailed description.

GetServiceAccountResponse
There is no detailed description.

GlobalExplanation
Global explanations containing the top most important features after training.

GoogleSheetsOptions
There is no detailed description.

HivePartitioningOptions
There is no detailed description.

IterationResult
Information about a single iteration of the training run.

Job
There is no detailed description.

JobCancelCall
Requests that a job be cancelled. This call will return immediately, and the client will need to poll for the job status to see if the cancel completed successfully. Cancelled jobs may still incur costs.

JobCancelResponse
There is no detailed description.

JobConfiguration
There is no detailed description.

JobConfigurationExtract
There is no detailed description.

JobConfigurationLoad
There is no detailed description.

JobConfigurationQuery
There is no detailed description.

JobConfigurationTableCopy
There is no detailed description.

JobDeleteCall
Requests that a job is deleted. This call will return when the job is deleted. This method is available in limited preview.

JobGetCall
Returns information about a specific job. Job information is available for a six month period after creation. Requires that you're the person who ran the job, or have the Is Owner project role.

JobGetQueryResultCall
Retrieves the results of a query job.

JobInsertCall
Starts a new asynchronous job. Requires the Can View project role.

JobList
There is no detailed description.

JobListCall
Lists all jobs that you started in the specified project. Job information is available for a six month period after creation. The job list is sorted in reverse chronological order, by job creation time. Requires the Can View project role, or the Is Owner project role if you set the allUsers property.

JobListJobs
List of jobs that were requested.

JobMethods
A builder providing access to all methods supported on job resources. It is not used directly, but through the Bigquery hub.

JobQueryCall
Runs a BigQuery SQL query synchronously and returns query results if the query completes within a specified timeout.

JobReference
There is no detailed description.

JobStatistics
There is no detailed description.

JobStatistics2
There is no detailed description.

JobStatistics2ReservationUsage
[Output-only] Job resource usage breakdown by reservation.

JobStatistics3
There is no detailed description.

JobStatistics4
There is no detailed description.

JobStatisticsReservationUsage
[Output-only] Job resource usage breakdown by reservation.

JobStatus
There is no detailed description.

JsonObject
Represents a single JSON object.

JsonValue
There is no detailed description.

ListModelsResponse
There is no detailed description.

ListRoutinesResponse
There is no detailed description.

ListRowAccessPoliciesResponse
Response message for the ListRowAccessPolicies method.

MaterializedViewDefinition
There is no detailed description.

Model
There is no detailed description.

ModelDefinition
There is no detailed description.

ModelDefinitionModelOptions
[Output-only, Beta] Model options used for the first training run. These options are immutable for subsequent training runs. Default values are used for any options not specified in the input query.

ModelDeleteCall
Deletes the model specified by modelId from the dataset.

ModelGetCall
Gets the specified model resource by model ID.

ModelListCall
Lists all models in the specified dataset. Requires the READER dataset role.

ModelMethods
A builder providing access to all methods supported on model resources. It is not used directly, but through the Bigquery hub.

ModelPatchCall
Patch specific fields in the specified model.

ModelReference
There is no detailed description.

MultiClassClassificationMetrics
Evaluation metrics for multi-class classification/classifier models.

ParquetOptions
There is no detailed description.

Policy
An Identity and Access Management (IAM) policy, which specifies access controls for Google Cloud resources. A Policy is a collection of bindings. A binding binds one or more members to a single role. Members can be user accounts, service accounts, Google groups, and domains (such as G Suite). A role is a named list of permissions; each role can be an IAM predefined role or a user-created custom role. For some types of Google Cloud resources, a binding can also specify a condition, which is a logical expression that allows access to a resource only if the expression evaluates to true. A condition can add constraints based on attributes of the request, the resource, or both. To learn which resources support conditions in their IAM policies, see the IAM documentation. JSON example: { "bindings": [ { "role": "roles/resourcemanager.organizationAdmin", "members": [ "user:mike@example.com", "group:admins@example.com", "domain:google.com", "serviceAccount:my-project-id@appspot.gserviceaccount.com" ] }, { "role": "roles/resourcemanager.organizationViewer", "members": [ "user:eve@example.com" ], "condition": { "title": "expirable access", "description": "Does not grant access after Sep 2020", "expression": "request.time < timestamp('2020-10-01T00:00:00.000Z')", } } ], "etag": "BwWWja0YfJA=", "version": 3 } YAML example: bindings: - members: - user:mike@example.com - group:admins@example.com - domain:google.com - serviceAccount:my-project-id@appspot.gserviceaccount.com role: roles/resourcemanager.organizationAdmin - members: - user:eve@example.com role: roles/resourcemanager.organizationViewer condition: title: expirable access description: Does not grant access after Sep 2020 expression: request.time < timestamp('2020-10-01T00:00:00.000Z') - etag: BwWWja0YfJA= - version: 3 For a description of IAM and its features, see the IAM documentation.

PrincipalComponentInfo
Principal component infos, used only for eigen decomposition based models, e.g., PCA. Ordered by explained_variance in the descending order.

ProjectGetServiceAccountCall
Returns the email address of the service account for your project used for interactions with Google Cloud KMS.

ProjectList
There is no detailed description.

ProjectListCall
Lists all projects to which you have been granted any project role.

ProjectListProjects
Projects to which you have at least READ access.

ProjectMethods
A builder providing access to all methods supported on project resources. It is not used directly, but through the Bigquery hub.

ProjectReference
There is no detailed description.

QueryParameter
There is no detailed description.

QueryParameterType
There is no detailed description.

QueryParameterTypeStructTypes
[Optional] The types of the fields of this struct, in order, if this is a struct.

QueryParameterValue
There is no detailed description.

QueryRequest
There is no detailed description.

QueryResponse
There is no detailed description.

QueryTimelineSample
There is no detailed description.

RangePartitioning
There is no detailed description.

RangePartitioningRange
[TrustedTester] [Required] Defines the ranges for range partitioning.

RankingMetrics
Evaluation metrics used by weighted-ALS models specified by feedback_type=implicit.

RegressionMetrics
Evaluation metrics for regression and explicit feedback type matrix factorization models.

Routine
A user-defined function or a stored procedure.

RoutineDeleteCall
Deletes the routine specified by routineId from the dataset.

RoutineGetCall
Gets the specified routine resource by routine ID.

RoutineInsertCall
Creates a new routine in the dataset.

RoutineListCall
Lists all routines in the specified dataset. Requires the READER dataset role.

RoutineMethods
A builder providing access to all methods supported on routine resources. It is not used directly, but through the Bigquery hub.

RoutineReference
There is no detailed description.

RoutineUpdateCall
Updates information in an existing routine. The update method replaces the entire Routine resource.

Row
A single row in the confusion matrix.

RowAccessPolicy
Represents access on a subset of rows on the specified table, defined by its filter predicate. Access to the subset of rows is controlled by its IAM policy.

RowAccessPolicyGetIamPolicyCall
Gets the access control policy for a resource. Returns an empty policy if the resource exists and does not have a policy set.

RowAccessPolicyListCall
Lists all row access policies on the specified table.

RowAccessPolicyMethods
A builder providing access to all methods supported on rowAccessPolicy resources. It is not used directly, but through the Bigquery hub.

RowAccessPolicyReference
There is no detailed description.

RowAccessPolicySetIamPolicyCall
Sets the access control policy on the specified resource. Replaces any existing policy. Can return NOT_FOUND, INVALID_ARGUMENT, and PERMISSION_DENIED errors.

RowAccessPolicyTestIamPermissionCall
Returns permissions that a caller has on the specified resource. If the resource does not exist, this will return an empty set of permissions, not a NOT_FOUND error. Note: This operation is designed to be used for building permission-aware UIs and command-line tools, not for authorization checking. This operation may "fail open" without warning.

RowLevelSecurityStatistics
There is no detailed description.

ScriptStackFrame
There is no detailed description.

ScriptStatistics
There is no detailed description.

SessionInfo
There is no detailed description.

SetIamPolicyRequest
Request message for SetIamPolicy method.

SnapshotDefinition
There is no detailed description.

StandardSqlDataType
The type of a variable, e.g., a function argument. Examples: INT64: {type_kind="INT64"} ARRAY: {type_kind="ARRAY", array_element_type="STRING"} STRUCT>: {type_kind="STRUCT", struct_type={fields=[ {name="x", type={type_kind="STRING"}}, {name="y", type={type_kind="ARRAY", array_element_type="DATE"}} ]}}

StandardSqlField
A field or a column.

StandardSqlStructType
There is no detailed description.

StandardSqlTableType
A table type

Streamingbuffer
There is no detailed description.

Table
There is no detailed description.

TableCell
There is no detailed description.

TableDataInsertAllRequest
There is no detailed description.

TableDataInsertAllRequestRows
The rows to insert.

TableDataInsertAllResponse
There is no detailed description.

TableDataInsertAllResponseInsertErrors
An array of errors for rows that were not inserted.

TableDataList
There is no detailed description.

TableDeleteCall
Deletes the table specified by tableId from the dataset. If the table contains data, all the data will be deleted.

TableFieldSchema
There is no detailed description.

TableFieldSchemaCategories
[Optional] The categories attached to this field, used for field-level access control.

TableFieldSchemaPolicyTags
There is no detailed description.

TableGetCall
Gets the specified table resource by table ID. This method does not return the data in the table, it only returns the table resource, which describes the structure of this table.

TableGetIamPolicyCall
Gets the access control policy for a resource. Returns an empty policy if the resource exists and does not have a policy set.

TableInsertCall
Creates a new, empty table in the dataset.

TableList
There is no detailed description.

TableListCall
Lists all tables in the specified dataset. Requires the READER dataset role.

TableListTables
Tables in the requested dataset.

TableListTablesView
Additional details for a view.

TableMethods
A builder providing access to all methods supported on table resources. It is not used directly, but through the Bigquery hub.

TablePatchCall
Updates information in an existing table. The update method replaces the entire table resource, whereas the patch method only replaces fields that are provided in the submitted table resource. This method supports patch semantics.

TableReference
There is no detailed description.

TableRow
There is no detailed description.

TableSchema
There is no detailed description.

TableSetIamPolicyCall
Sets the access control policy on the specified resource. Replaces any existing policy. Can return NOT_FOUND, INVALID_ARGUMENT, and PERMISSION_DENIED errors.

TableTestIamPermissionCall
Returns permissions that a caller has on the specified resource. If the resource does not exist, this will return an empty set of permissions, not a NOT_FOUND error. Note: This operation is designed to be used for building permission-aware UIs and command-line tools, not for authorization checking. This operation may "fail open" without warning.

TableUpdateCall
Updates information in an existing table. The update method replaces the entire table resource, whereas the patch method only replaces fields that are provided in the submitted table resource.

TabledataInsertAllCall
Streams data into BigQuery one record at a time without needing to run a load job. Requires the WRITER dataset role.

TabledataListCall
Retrieves table data from a specified set of rows. Requires the READER dataset role.

TabledataMethods
A builder providing access to all methods supported on tabledata resources. It is not used directly, but through the Bigquery hub.

TestIamPermissionsRequest
Request message for TestIamPermissions method.

TestIamPermissionsResponse
Response message for TestIamPermissions method.

TimePartitioning
There is no detailed description.

TrainingOptions
Options used in model training.

TrainingRun
Information about a single training query run for the model.

TransactionInfo
There is no detailed description.

UserDefinedFunctionResource
This is used for defining User Defined Function (UDF) resources only when using legacy SQL. Users of Standard SQL should leverage either DDL (e.g. CREATE [TEMPORARY] FUNCTION ... ) or the Routines API to define UDF resources. For additional information on migrating, see: https://cloud.google.com/bigquery/docs/reference/standard-sql/migrating-from-legacy-sql#differences_in_user-defined_javascript_functions

ViewDefinition
There is no detailed description.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.