Struct google_bigquery2::api::JobConfigurationQuery

pub struct JobConfigurationQuery {
    pub allow_large_results: Option<bool>,
    pub clustering: Option<Clustering>,
    pub connection_properties: Option<Vec<ConnectionProperty>>,
    pub create_disposition: Option<String>,
    pub create_session: Option<bool>,
    pub default_dataset: Option<DatasetReference>,
    pub destination_encryption_configuration: Option<EncryptionConfiguration>,
    pub destination_table: Option<TableReference>,
    pub flatten_results: Option<bool>,
    pub maximum_billing_tier: Option<i32>,
    pub maximum_bytes_billed: Option<String>,
    pub parameter_mode: Option<String>,
    pub preserve_nulls: Option<bool>,
    pub priority: Option<String>,
    pub query: Option<String>,
    pub query_parameters: Option<Vec<QueryParameter>>,
    pub range_partitioning: Option<RangePartitioning>,
    pub schema_update_options: Option<Vec<String>>,
    pub table_definitions: Option<HashMap<String, ExternalDataConfiguration>>,
    pub time_partitioning: Option<TimePartitioning>,
    pub use_legacy_sql: Option<bool>,
    pub use_query_cache: Option<bool>,
    pub user_defined_function_resources: Option<Vec<UserDefinedFunctionResource>>,
    pub write_disposition: Option<String>,
}

There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

Fields
allow_large_results: Option<bool>

[Optional] If true and query uses legacy SQL dialect, allows the query to produce arbitrarily large result tables at a slight cost in performance. Requires destinationTable to be set. For standard SQL queries, this flag is ignored and large results are always allowed. However, you must still set destinationTable when result size exceeds the allowed maximum response size.

clustering: Option<Clustering>

[Beta] Clustering specification for the destination table. Must be specified with time-based partitioning, data in the table will be first partitioned and subsequently clustered.

connection_properties: Option<Vec<ConnectionProperty>>

Connection properties.

create_disposition: Option<String>

[Optional] Specifies whether the job is allowed to create new tables. The following values are supported: CREATE_IF_NEEDED: If the table does not exist, BigQuery creates the table. CREATE_NEVER: The table must already exist. If it does not, a 'notFound' error is returned in the job result. The default value is CREATE_IF_NEEDED. Creation, truncation and append actions occur as one atomic update upon job completion.

create_session: Option<bool>

If true, creates a new session, where session id will be a server generated random id. If false, runs query with an existing session_id passed in ConnectionProperty, otherwise runs query in non-session mode.

default_dataset: Option<DatasetReference>

[Optional] Specifies the default dataset to use for unqualified table names in the query. Note that this does not alter behavior of unqualified dataset names.

destination_encryption_configuration: Option<EncryptionConfiguration>

Custom encryption configuration (e.g., Cloud KMS keys).

destination_table: Option<TableReference>

[Optional] Describes the table where the query results should be stored. If not present, a new table will be created to store the results. This property must be set for large results that exceed the maximum response size.

flatten_results: Option<bool>

[Optional] If true and query uses legacy SQL dialect, flattens all nested and repeated fields in the query results. allowLargeResults must be true if this is set to false. For standard SQL queries, this flag is ignored and results are never flattened.

maximum_billing_tier: Option<i32>

[Optional] Limits the billing tier for this job. Queries that have resource usage beyond this tier will fail (without incurring a charge). If unspecified, this will be set to your project default.

maximum_bytes_billed: Option<String>

[Optional] Limits the bytes billed for this job. Queries that will have bytes billed beyond this limit will fail (without incurring a charge). If unspecified, this will be set to your project default.

parameter_mode: Option<String>

Standard SQL only. Set to POSITIONAL to use positional (?) query parameters or to NAMED to use named (@myparam) query parameters in this query.

preserve_nulls: Option<bool>

[Deprecated] This property is deprecated.

priority: Option<String>

[Optional] Specifies a priority for the query. Possible values include INTERACTIVE and BATCH. The default value is INTERACTIVE.

query: Option<String>

[Required] SQL query text to execute. The useLegacySql field can be used to indicate whether the query uses legacy SQL or standard SQL.

query_parameters: Option<Vec<QueryParameter>>

Query parameters for standard SQL queries.

range_partitioning: Option<RangePartitioning>

[TrustedTester] Range partitioning specification for this table. Only one of timePartitioning and rangePartitioning should be specified.

schema_update_options: Option<Vec<String>>

Allows the schema of the destination table to be updated as a side effect of the query job. Schema update options are supported in two cases: when writeDisposition is WRITE_APPEND; when writeDisposition is WRITE_TRUNCATE and the destination table is a partition of a table, specified by partition decorators. For normal tables, WRITE_TRUNCATE will always overwrite the schema. One or more of the following values are specified: ALLOW_FIELD_ADDITION: allow adding a nullable field to the schema. ALLOW_FIELD_RELAXATION: allow relaxing a required field in the original schema to nullable.

table_definitions: Option<HashMap<String, ExternalDataConfiguration>>

[Optional] If querying an external data source outside of BigQuery, describes the data format, location and other properties of the data source. By defining these properties, the data source can then be queried as if it were a standard BigQuery table.

time_partitioning: Option<TimePartitioning>

Time-based partitioning specification for the destination table. Only one of timePartitioning and rangePartitioning should be specified.

use_legacy_sql: Option<bool>

Specifies whether to use BigQuery's legacy SQL dialect for this query. The default value is true. If set to false, the query will use BigQuery's standard SQL: https://cloud.google.com/bigquery/sql-reference/ When useLegacySql is set to false, the value of flattenResults is ignored; query will be run as if flattenResults is false.

use_query_cache: Option<bool>

[Optional] Whether to look for the result in the query cache. The query cache is a best-effort cache that will be flushed whenever tables in the query are modified. Moreover, the query cache is only available when a query does not have a destination table specified. The default value is true.

user_defined_function_resources: Option<Vec<UserDefinedFunctionResource>>

Describes user-defined function resources used in the query.

write_disposition: Option<String>

[Optional] Specifies the action that occurs if the destination table already exists. The following values are supported: WRITE_TRUNCATE: If the table already exists, BigQuery overwrites the table data and uses the schema from the query result. WRITE_APPEND: If the table already exists, BigQuery appends the data to the table. WRITE_EMPTY: If the table already exists and contains data, a 'duplicate' error is returned in the job result. The default value is WRITE_EMPTY. Each action is atomic and only occurs if BigQuery is able to complete the job successfully. Creation, truncation and append actions occur as one atomic update upon job completion.