Struct google_bigquery2::api::JobStatistics

pub struct JobStatistics {
    pub completion_ratio: Option<f64>,
    pub creation_time: Option<String>,
    pub end_time: Option<String>,
    pub extract: Option<JobStatistics4>,
    pub load: Option<JobStatistics3>,
    pub num_child_jobs: Option<String>,
    pub parent_job_id: Option<String>,
    pub query: Option<JobStatistics2>,
    pub quota_deferments: Option<Vec<String>>,
    pub reservation_usage: Option<Vec<JobStatisticsReservationUsage>>,
    pub reservation_id: Option<String>,
    pub row_level_security_statistics: Option<RowLevelSecurityStatistics>,
    pub script_statistics: Option<ScriptStatistics>,
    pub session_info_template: Option<SessionInfo>,
    pub start_time: Option<String>,
    pub total_bytes_processed: Option<String>,
    pub total_slot_ms: Option<String>,
    pub transaction_info_template: Option<TransactionInfo>,
}

There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

Fields
completion_ratio: Option<f64>

[TrustedTester] [Output-only] Job progress (0.0 -> 1.0) for LOAD and EXTRACT jobs.

creation_time: Option<String>

[Output-only] Creation time of this job, in milliseconds since the epoch. This field will be present on all jobs.

end_time: Option<String>

[Output-only] End time of this job, in milliseconds since the epoch. This field will be present whenever a job is in the DONE state.

extract: Option<JobStatistics4>

[Output-only] Statistics for an extract job.

load: Option<JobStatistics3>

[Output-only] Statistics for a load job.

num_child_jobs: Option<String>

[Output-only] Number of child jobs executed.

parent_job_id: Option<String>

[Output-only] If this is a child job, the id of the parent.

query: Option<JobStatistics2>

[Output-only] Statistics for a query job.

quota_deferments: Option<Vec<String>>

[Output-only] Quotas which delayed this job's start time.

reservation_usage: Option<Vec<JobStatisticsReservationUsage>>

[Output-only] Job resource usage breakdown by reservation.

reservation_id: Option<String>

[Output-only] Name of the primary reservation assigned to this job. Note that this could be different than reservations reported in the reservation usage field if parent reservations were used to execute this job.

row_level_security_statistics: Option<RowLevelSecurityStatistics>

[Output-only] [Preview] Statistics for row-level security. Present only for query and extract jobs.

script_statistics: Option<ScriptStatistics>

[Output-only] Statistics for a child job of a script.

session_info_template: Option<SessionInfo>

[Output-only] [Preview] Information of the session if this job is part of one.

start_time: Option<String>

[Output-only] Start time of this job, in milliseconds since the epoch. This field will be present when the job transitions from the PENDING state to either RUNNING or DONE.

total_bytes_processed: Option<String>

[Output-only] [Deprecated] Use the bytes processed in the query statistics instead.

total_slot_ms: Option<String>

[Output-only] Slot-milliseconds for the job.

transaction_info_template: Option<TransactionInfo>

[Output-only] [Alpha] Information of the multi-statement transaction if this job is part of one.