Struct google_bigquery2::api::QueryRequest

pub struct QueryRequest {
    pub connection_properties: Option<Vec<ConnectionProperty>>,
    pub create_session: Option<bool>,
    pub default_dataset: Option<DatasetReference>,
    pub dry_run: Option<bool>,
    pub kind: Option<String>,
    pub labels: Option<HashMap<String, String>>,
    pub location: Option<String>,
    pub max_results: Option<u32>,
    pub maximum_bytes_billed: Option<String>,
    pub parameter_mode: Option<String>,
    pub preserve_nulls: Option<bool>,
    pub query: Option<String>,
    pub query_parameters: Option<Vec<QueryParameter>>,
    pub request_id: Option<String>,
    pub timeout_ms: Option<u32>,
    pub use_legacy_sql: Option<bool>,
    pub use_query_cache: Option<bool>,
}

There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

query jobs (request)

Fields
connection_properties: Option<Vec<ConnectionProperty>>

Connection properties.

create_session: Option<bool>

If true, creates a new session, where session id will be a server generated random id. If false, runs query with an existing session_id passed in ConnectionProperty, otherwise runs query in non-session mode.

default_dataset: Option<DatasetReference>

[Optional] Specifies the default datasetId and projectId to assume for any unqualified table names in the query. If not set, all table names in the query string must be qualified in the format 'datasetId.tableId'.

dry_run: Option<bool>

[Optional] If set to true, BigQuery doesn't run the job. Instead, if the query is valid, BigQuery returns statistics about the job such as how many bytes would be processed. If the query is invalid, an error returns. The default value is false.

kind: Option<String>

The resource type of the request.

labels: Option<HashMap<String, String>>

The labels associated with this job. You can use these to organize and group your jobs. Label keys and values can be no longer than 63 characters, can only contain lowercase letters, numeric characters, underscores and dashes. International characters are allowed. Label values are optional. Label keys must start with a letter and each label in the list must have a different key.

location: Option<String>

The geographic location where the job should run. See details at https://cloud.google.com/bigquery/docs/locations#specifying_your_location.

max_results: Option<u32>

[Optional] The maximum number of rows of data to return per page of results. Setting this flag to a small value such as 1000 and then paging through results might improve reliability when the query result set is large. In addition to this limit, responses are also limited to 10 MB. By default, there is no maximum row count, and only the byte limit applies.

maximum_bytes_billed: Option<String>

[Optional] Limits the bytes billed for this job. Queries that will have bytes billed beyond this limit will fail (without incurring a charge). If unspecified, this will be set to your project default.

parameter_mode: Option<String>

Standard SQL only. Set to POSITIONAL to use positional (?) query parameters or to NAMED to use named (@myparam) query parameters in this query.

preserve_nulls: Option<bool>

[Deprecated] This property is deprecated.

query: Option<String>

[Required] A query string, following the BigQuery query syntax, of the query to execute. Example: "SELECT count(f1) FROM [myProjectId:myDatasetId.myTableId]".

query_parameters: Option<Vec<QueryParameter>>

Query parameters for Standard SQL queries.

request_id: Option<String>

A unique user provided identifier to ensure idempotent behavior for queries. Note that this is different from the job_id. It has the following properties: 1. It is case-sensitive, limited to up to 36 ASCII characters. A UUID is recommended. 2. Read only queries can ignore this token since they are nullipotent by definition. 3. For the purposes of idempotency ensured by the request_id, a request is considered duplicate of another only if they have the same request_id and are actually duplicates. When determining whether a request is a duplicate of the previous request, all parameters in the request that may affect the behavior are considered. For example, query, connection_properties, query_parameters, use_legacy_sql are parameters that affect the result and are considered when determining whether a request is a duplicate, but properties like timeout_ms don't affect the result and are thus not considered. Dry run query requests are never considered duplicate of another request. 4. When a duplicate mutating query request is detected, it returns: a. the results of the mutation if it completes successfully within the timeout. b. the running operation if it is still in progress at the end of the timeout. 5. Its lifetime is limited to 15 minutes. In other words, if two requests are sent with the same request_id, but more than 15 minutes apart, idempotency is not guaranteed.

timeout_ms: Option<u32>

[Optional] How long to wait for the query to complete, in milliseconds, before the request times out and returns. Note that this is only a timeout for the request, not the query. If the query takes longer to run than the timeout value, the call returns without any results and with the 'jobComplete' flag set to false. You can call GetQueryResults() to wait for the query to complete and read the results. The default value is 10000 milliseconds (10 seconds).

use_legacy_sql: Option<bool>

Specifies whether to use BigQuery's legacy SQL dialect for this query. The default value is true. If set to false, the query will use BigQuery's standard SQL: https://cloud.google.com/bigquery/sql-reference/ When useLegacySql is set to false, the value of flattenResults is ignored; query will be run as if flattenResults is false.

use_query_cache: Option<bool>

[Optional] Whether to look for the result in the query cache. The query cache is a best-effort cache that will be flushed whenever tables in the query are modified. The default value is true.