Struct google_bigquery2::api::TrainingOptions

pub struct TrainingOptions {
    pub auto_arima: Option<bool>,
    pub auto_arima_max_order: Option<String>,
    pub batch_size: Option<String>,
    pub data_frequency: Option<String>,
    pub data_split_column: Option<String>,
    pub data_split_eval_fraction: Option<f64>,
    pub data_split_method: Option<String>,
    pub distance_type: Option<String>,
    pub dropout: Option<f64>,
    pub early_stop: Option<bool>,
    pub feedback_type: Option<String>,
    pub hidden_units: Option<Vec<String>>,
    pub holiday_region: Option<String>,
    pub horizon: Option<String>,
    pub include_drift: Option<bool>,
    pub initial_learn_rate: Option<f64>,
    pub input_label_columns: Option<Vec<String>>,
    pub item_column: Option<String>,
    pub kmeans_initialization_column: Option<String>,
    pub kmeans_initialization_method: Option<String>,
    pub l1_regularization: Option<f64>,
    pub l2_regularization: Option<f64>,
    pub label_class_weights: Option<HashMap<String, f64>>,
    pub learn_rate: Option<f64>,
    pub learn_rate_strategy: Option<String>,
    pub loss_type: Option<String>,
    pub max_iterations: Option<String>,
    pub max_tree_depth: Option<String>,
    pub min_relative_progress: Option<f64>,
    pub min_split_loss: Option<f64>,
    pub model_uri: Option<String>,
    pub non_seasonal_order: Option<ArimaOrder>,
    pub num_clusters: Option<String>,
    pub num_factors: Option<String>,
    pub optimization_strategy: Option<String>,
    pub preserve_input_structs: Option<bool>,
    pub subsample: Option<f64>,
    pub time_series_data_column: Option<String>,
    pub time_series_id_column: Option<String>,
    pub time_series_timestamp_column: Option<String>,
    pub user_column: Option<String>,
    pub wals_alpha: Option<f64>,
    pub warm_start: Option<bool>,
}

Options used in model training.

This type is not used in any activity, and only used as part of another schema.

Fields
auto_arima: Option<bool>

Whether to enable auto ARIMA or not.

auto_arima_max_order: Option<String>

The max value of non-seasonal p and q.

batch_size: Option<String>

Batch size for dnn models.

data_frequency: Option<String>

The data frequency of a time series.

data_split_column: Option<String>

The column to split data with. This column won't be used as a feature. 1. When data_split_method is CUSTOM, the corresponding column should be boolean. The rows with true value tag are eval data, and the false are training data. 2. When data_split_method is SEQ, the first DATA_SPLIT_EVAL_FRACTION rows (from smallest to largest) in the corresponding column are used as training data, and the rest are eval data. It respects the order in Orderable data types: https://cloud.google.com/bigquery/docs/reference/standard-sql/data-types#data-type-properties

data_split_eval_fraction: Option<f64>

The fraction of evaluation data over the whole input data. The rest of data will be used as training data. The format should be double. Accurate to two decimal places. Default value is 0.2.

data_split_method: Option<String>

The data split type for training and evaluation, e.g. RANDOM.

distance_type: Option<String>

Distance type for clustering models.

dropout: Option<f64>

Dropout probability for dnn models.

early_stop: Option<bool>

Whether to stop early when the loss doesn't improve significantly any more (compared to min_relative_progress). Used only for iterative training algorithms.

feedback_type: Option<String>

Feedback type that specifies which algorithm to run for matrix factorization.

hidden_units: Option<Vec<String>>

Hidden units for dnn models.

holiday_region: Option<String>

The geographical region based on which the holidays are considered in time series modeling. If a valid value is specified, then holiday effects modeling is enabled.

horizon: Option<String>

The number of periods ahead that need to be forecasted.

include_drift: Option<bool>

Include drift when fitting an ARIMA model.

initial_learn_rate: Option<f64>

Specifies the initial learning rate for the line search learn rate strategy.

input_label_columns: Option<Vec<String>>

Name of input label columns in training data.

item_column: Option<String>

Item column specified for matrix factorization models.

kmeans_initialization_column: Option<String>

The column used to provide the initial centroids for kmeans algorithm when kmeans_initialization_method is CUSTOM.

kmeans_initialization_method: Option<String>

The method used to initialize the centroids for kmeans algorithm.

l1_regularization: Option<f64>

L1 regularization coefficient.

l2_regularization: Option<f64>

L2 regularization coefficient.

label_class_weights: Option<HashMap<String, f64>>

Weights associated with each label class, for rebalancing the training data. Only applicable for classification models.

learn_rate: Option<f64>

Learning rate in training. Used only for iterative training algorithms.

learn_rate_strategy: Option<String>

The strategy to determine learn rate for the current iteration.

loss_type: Option<String>

Type of loss function used during training run.

max_iterations: Option<String>

The maximum number of iterations in training. Used only for iterative training algorithms.

max_tree_depth: Option<String>

Maximum depth of a tree for boosted tree models.

min_relative_progress: Option<f64>

When early_stop is true, stops training when accuracy improvement is less than 'min_relative_progress'. Used only for iterative training algorithms.

min_split_loss: Option<f64>

Minimum split loss for boosted tree models.

model_uri: Option<String>

Google Cloud Storage URI from which the model was imported. Only applicable for imported models.

non_seasonal_order: Option<ArimaOrder>

A specification of the non-seasonal part of the ARIMA model: the three components (p, d, q) are the AR order, the degree of differencing, and the MA order.

num_clusters: Option<String>

Number of clusters for clustering models.

num_factors: Option<String>

Num factors specified for matrix factorization models.

optimization_strategy: Option<String>

Optimization strategy for training linear regression models.

preserve_input_structs: Option<bool>

Whether to preserve the input structs in output feature names. Suppose there is a struct A with field b. When false (default), the output feature name is A_b. When true, the output feature name is A.b.

subsample: Option<f64>

Subsample fraction of the training data to grow tree to prevent overfitting for boosted tree models.

time_series_data_column: Option<String>

Column to be designated as time series data for ARIMA model.

time_series_id_column: Option<String>

The time series id column that was used during ARIMA model training.

time_series_timestamp_column: Option<String>

Column to be designated as time series timestamp for ARIMA model.

user_column: Option<String>

User column specified for matrix factorization models.

wals_alpha: Option<f64>

Hyperparameter for matrix factoration when implicit feedback type is specified.

warm_start: Option<bool>

Whether to train a model from the last checkpoint.