Struct google_bigquery2::CsvOptions

pub struct CsvOptions {
    pub allow_jagged_rows: Option<bool>,
    pub skip_leading_rows: Option<i32>,
    pub field_delimiter: Option<String>,
    pub encoding: Option<String>,
    pub quote: Option<String>,
    pub allow_quoted_newlines: Option<bool>,
}

There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

Fields

allow_jagged_rows

[Optional] Indicates if BigQuery should accept rows that are missing trailing optional columns. If true, BigQuery treats missing trailing columns as null values. If false, records with missing trailing columns are treated as bad records, and if there are too many bad records, an invalid error is returned in the job result. The default value is false.

skip_leading_rows

[Optional] The number of rows at the top of a CSV file that BigQuery will skip when reading the data. The default value is 0. This property is useful if you have header rows in the file that should be skipped.

field_delimiter

[Optional] The separator for fields in a CSV file. BigQuery converts the string to ISO-8859-1 encoding, and then uses the first byte of the encoded string to split the data in its raw, binary state. BigQuery also supports the escape sequence "\t" to specify a tab separator. The default value is a comma (',').

encoding

[Optional] The character encoding of the data. The supported values are UTF-8 or ISO-8859-1. The default value is UTF-8. BigQuery decodes the data after the raw, binary data has been split using the values of the quote and fieldDelimiter properties.

quote

[Optional] The value that is used to quote data sections in a CSV file. BigQuery converts the string to ISO-8859-1 encoding, and then uses the first byte of the encoded string to split the data in its raw, binary state. The default value is a double-quote ('"'). If your data does not contain quoted sections, set the property value to an empty string. If your data contains quoted newline characters, you must also set the allowQuotedNewlines property to true.

allow_quoted_newlines

[Optional] Indicates if BigQuery should allow quoted data sections that contain newline characters in a CSV file. The default value is false.