Struct google_bigquery2::DatasetList

pub struct DatasetList {
    pub next_page_token: Option<String>,
    pub kind: Option<String>,
    pub datasets: Option<Vec<DatasetListDatasets>>,
    pub etag: Option<String>,
}

There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list datasets (response)

Fields

next_page_token

A token that can be used to request the next results page. This property is omitted on the final results page.

kind

The list type. This property always returns the value "bigquery#datasetList".

datasets

An array of the dataset resources in the project. Each resource contains basic information. For full information about a particular dataset resource, use the Datasets: get method. This property is omitted when there are no datasets in the project.

etag

A hash value of the results page. You can use this property to determine if the page has changed since the last request.