Struct google_bigquery2::DatasetMethods

pub struct DatasetMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on dataset resources. It is not used directly, but through the Bigquery hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_bigquery2 as bigquery2;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use bigquery2::Bigquery;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Bigquery::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `delete(...)`, `get(...)`, `insert(...)`, `list(...)`, `patch(...)` and `update(...)`
// to build up your call.
let rb = hub.datasets();

Methods

impl<'a, C, A> DatasetMethods<'a, C, A>

fn patch(&self, request: Dataset, project_id: &str, dataset_id: &str) -> DatasetPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates information in an existing dataset. The update method replaces the entire dataset resource, whereas the patch method only replaces fields that are provided in the submitted dataset resource. This method supports patch semantics.

Arguments

request - No description provided.

projectId - Project ID of the dataset being updated

datasetId - Dataset ID of the dataset being updated

fn get(&self, project_id: &str, dataset_id: &str) -> DatasetGetCall<'a, C, A>

Create a builder to help you perform the following task:

Returns the dataset specified by datasetID.

Arguments

projectId - Project ID of the requested dataset

datasetId - Dataset ID of the requested dataset

fn list(&self, project_id: &str) -> DatasetListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists all datasets in the specified project to which you have been granted the READER dataset role.

Arguments

projectId - Project ID of the datasets to be listed

fn update(&self, request: Dataset, project_id: &str, dataset_id: &str) -> DatasetUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates information in an existing dataset. The update method replaces the entire dataset resource, whereas the patch method only replaces fields that are provided in the submitted dataset resource.

Arguments

request - No description provided.

projectId - Project ID of the dataset being updated

datasetId - Dataset ID of the dataset being updated

fn delete(&self, project_id: &str, dataset_id: &str) -> DatasetDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes the dataset specified by the datasetId value. Before you can delete a dataset, you must delete all its tables, either manually or by specifying deleteContents. Immediately after deletion, you can create another dataset with the same name.

Arguments

projectId - Project ID of the dataset being deleted

datasetId - Dataset ID of dataset being deleted

fn insert(&self, request: Dataset, project_id: &str) -> DatasetInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a new empty dataset.

Arguments

request - No description provided.

projectId - Project ID of the new dataset