Struct google_bigquery2::ExplainQueryStage

pub struct ExplainQueryStage {
    pub status: Option<String>,
    pub shuffle_output_bytes: Option<String>,
    pub compute_ms_max: Option<String>,
    pub input_stages: Option<Vec<String>>,
    pub compute_ms_avg: Option<String>,
    pub start_ms: Option<String>,
    pub shuffle_output_bytes_spilled: Option<String>,
    pub wait_ms_avg: Option<String>,
    pub write_ms_avg: Option<String>,
    pub read_ms_avg: Option<String>,
    pub wait_ratio_avg: Option<f64>,
    pub id: Option<String>,
    pub end_ms: Option<String>,
    pub compute_ratio_max: Option<f64>,
    pub name: Option<String>,
    pub wait_ms_max: Option<String>,
    pub read_ms_max: Option<String>,
    pub records_written: Option<String>,
    pub wait_ratio_max: Option<f64>,
    pub read_ratio_avg: Option<f64>,
    pub parallel_inputs: Option<String>,
    pub read_ratio_max: Option<f64>,
    pub steps: Option<Vec<ExplainQueryStep>>,
    pub records_read: Option<String>,
    pub compute_ratio_avg: Option<f64>,
    pub completed_parallel_inputs: Option<String>,
    pub write_ratio_max: Option<f64>,
    pub write_ms_max: Option<String>,
    pub write_ratio_avg: Option<f64>,
}

There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

Fields
status: Option<String>

Current status for the stage.

shuffle_output_bytes: Option<String>

Total number of bytes written to shuffle.

compute_ms_max: Option<String>

Milliseconds the slowest shard spent on CPU-bound tasks.

input_stages: Option<Vec<String>>

IDs for stages that are inputs to this stage.

compute_ms_avg: Option<String>

Milliseconds the average shard spent on CPU-bound tasks.

start_ms: Option<String>

Stage start time represented as milliseconds since epoch.

shuffle_output_bytes_spilled: Option<String>

Total number of bytes written to shuffle and spilled to disk.

wait_ms_avg: Option<String>

Milliseconds the average shard spent waiting to be scheduled.

write_ms_avg: Option<String>

Milliseconds the average shard spent on writing output.

read_ms_avg: Option<String>

Milliseconds the average shard spent reading input.

wait_ratio_avg: Option<f64>

Relative amount of time the average shard spent waiting to be scheduled.

id: Option<String>

Unique ID for stage within plan.

end_ms: Option<String>

Stage end time represented as milliseconds since epoch.

compute_ratio_max: Option<f64>

Relative amount of time the slowest shard spent on CPU-bound tasks.

name: Option<String>

Human-readable name for stage.

wait_ms_max: Option<String>

Milliseconds the slowest shard spent waiting to be scheduled.

read_ms_max: Option<String>

Milliseconds the slowest shard spent reading input.

records_written: Option<String>

Number of records written by the stage.

wait_ratio_max: Option<f64>

Relative amount of time the slowest shard spent waiting to be scheduled.

read_ratio_avg: Option<f64>

Relative amount of time the average shard spent reading input.

parallel_inputs: Option<String>

Number of parallel input segments to be processed.

read_ratio_max: Option<f64>

Relative amount of time the slowest shard spent reading input.

steps: Option<Vec<ExplainQueryStep>>

List of operations within the stage in dependency order (approximately chronological).

records_read: Option<String>

Number of records read into the stage.

compute_ratio_avg: Option<f64>

Relative amount of time the average shard spent on CPU-bound tasks.

completed_parallel_inputs: Option<String>

Number of parallel input segments completed.

write_ratio_max: Option<f64>

Relative amount of time the slowest shard spent on writing output.

write_ms_max: Option<String>

Milliseconds the slowest shard spent on writing output.

write_ratio_avg: Option<f64>

Relative amount of time the average shard spent on writing output.