Struct google_bigquery2::ExternalDataConfiguration

pub struct ExternalDataConfiguration {
    pub compression: Option<String>,
    pub csv_options: Option<CsvOptions>,
    pub autodetect: Option<bool>,
    pub max_bad_records: Option<i32>,
    pub ignore_unknown_values: Option<bool>,
    pub source_uris: Option<Vec<String>>,
    pub bigtable_options: Option<BigtableOptions>,
    pub source_format: Option<String>,
    pub google_sheets_options: Option<GoogleSheetsOptions>,
    pub schema: Option<TableSchema>,
}

There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

Fields
compression: Option<String>

[Optional] The compression type of the data source. Possible values include GZIP and NONE. The default value is NONE. This setting is ignored for Google Cloud Bigtable, Google Cloud Datastore backups and Avro formats.

csv_options: Option<CsvOptions>

Additional properties to set if sourceFormat is set to CSV.

autodetect: Option<bool>

[Experimental] Try to detect schema and format options automatically. Any option specified explicitly will be honored.

max_bad_records: Option<i32>

[Optional] The maximum number of bad records that BigQuery can ignore when reading data. If the number of bad records exceeds this value, an invalid error is returned in the job result. The default value is 0, which requires that all records are valid. This setting is ignored for Google Cloud Bigtable, Google Cloud Datastore backups and Avro formats.

ignore_unknown_values: Option<bool>

[Optional] Indicates if BigQuery should allow extra values that are not represented in the table schema. If true, the extra values are ignored. If false, records with extra columns are treated as bad records, and if there are too many bad records, an invalid error is returned in the job result. The default value is false. The sourceFormat property determines what BigQuery treats as an extra value: CSV: Trailing columns JSON: Named values that don't match any column names Google Cloud Bigtable: This setting is ignored. Google Cloud Datastore backups: This setting is ignored. Avro: This setting is ignored.

source_uris: Option<Vec<String>>

[Required] The fully-qualified URIs that point to your data in Google Cloud. For Google Cloud Storage URIs: Each URI can contain one '*' wildcard character and it must come after the 'bucket' name. Size limits related to load jobs apply to external data sources. For Google Cloud Bigtable URIs: Exactly one URI can be specified and it has be a fully specified and valid HTTPS URL for a Google Cloud Bigtable table. For Google Cloud Datastore backups, exactly one URI can be specified, and it must end with '.backup_info'. Also, the '*' wildcard character is not allowed.

bigtable_options: Option<BigtableOptions>

[Optional] Additional options if sourceFormat is set to BIGTABLE.

source_format: Option<String>

[Required] The data format. For CSV files, specify "CSV". For Google sheets, specify "GOOGLE_SHEETS". For newline-delimited JSON, specify "NEWLINE_DELIMITED_JSON". For Avro files, specify "AVRO". For Google Cloud Datastore backups, specify "DATASTORE_BACKUP". [Experimental] For Google Cloud Bigtable, specify "BIGTABLE". Please note that reading from Google Cloud Bigtable is experimental and has to be enabled for your project. Please contact Google Cloud Support to enable this for your project.

google_sheets_options: Option<GoogleSheetsOptions>

[Optional] Additional options if sourceFormat is set to GOOGLE_SHEETS.

schema: Option<TableSchema>

[Optional] The schema for the data. Schema is required for CSV and JSON formats. Schema is disallowed for Google Cloud Bigtable, Cloud Datastore backups, and Avro formats.