Struct google_bigquery2::Job

pub struct Job {
    pub status: Option<JobStatus>,
    pub kind: Option<String>,
    pub statistics: Option<JobStatistics>,
    pub job_reference: Option<JobReference>,
    pub etag: Option<String>,
    pub self_link: Option<String>,
    pub configuration: Option<JobConfiguration>,
    pub id: Option<String>,
    pub user_email: Option<String>,
}

There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

insert jobs (request|response)

query jobs (none)

list jobs (none)

get query results jobs (none)

get jobs (response)

Fields

status

[Output-only] The status of this job. Examine this value when polling an asynchronous job to see if the job is complete.

kind

[Output-only] The type of the resource.

statistics

[Output-only] Information about the job, including starting time and ending time of the job.

job_reference

[Optional] Reference describing the unique-per-user name of the job.

etag

[Output-only] A hash of this resource.

self_link

[Output-only] A URL that can be used to access this resource again.

configuration

[Required] Describes the job configuration.

id

[Output-only] Opaque ID field of the job

user_email

[Output-only] Email address of the user who ran the job.