Struct google-bigquery2::JobConfigurationExtract

pub struct JobConfigurationExtract {
    pub destination_uri: String,
    pub destination_format: String,
    pub compression: String,
    pub destination_uris: Vec<String>,
    pub print_header: bool,
    pub field_delimiter: String,
    pub source_table: TableReference,
}

There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

Fields

destination_uri

[Pick one] DEPRECATED: Use destinationUris instead, passing only one URI as necessary. The fully-qualified Google Cloud Storage URI where the extracted table should be written.

destination_format

[Optional] The exported file format. Possible values include CSV, NEWLINE_DELIMITED_JSON and AVRO. The default value is CSV. Tables with nested or repeated fields cannot be exported as CSV.

compression

[Optional] The compression type to use for exported files. Possible values include GZIP and NONE. The default value is NONE.

destination_uris

[Pick one] A list of fully-qualified Google Cloud Storage URIs where the extracted table should be written.

print_header

[Optional] Whether to print out a header row in the results. Default is true.

field_delimiter

[Optional] Delimiter to use between fields in the exported data. Default is ','

source_table

[Required] A reference to the table being exported.