Struct google-bigquery2::JobConfigurationLink

pub struct JobConfigurationLink {
    pub create_disposition: Option<String>,
    pub write_disposition: Option<String>,
    pub destination_table: Option<TableReference>,
    pub source_uri: Vec<String>,
}

There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

Fields

create_disposition

[Optional] Specifies whether the job is allowed to create new tables. The following values are supported: CREATE_IF_NEEDED: If the table does not exist, BigQuery creates the table. CREATE_NEVER: The table must already exist. If it does not, a 'notFound' error is returned in the job result. The default value is CREATE_IF_NEEDED. Creation, truncation and append actions occur as one atomic update upon job completion.

write_disposition

[Optional] Specifies the action that occurs if the destination table already exists. The following values are supported: WRITE_TRUNCATE: If the table already exists, BigQuery overwrites the table data. WRITE_APPEND: If the table already exists, BigQuery appends the data to the table. WRITE_EMPTY: If the table already exists and contains data, a 'duplicate' error is returned in the job result. The default value is WRITE_EMPTY. Each action is atomic and only occurs if BigQuery is able to complete the job successfully. Creation, truncation and append actions occur as one atomic update upon job completion.

destination_table

[Required] The destination table of the link job.

source_uri

[Required] URI of source table to link.