Struct google_bigquery2::JobConfigurationLoad

pub struct JobConfigurationLoad {
    pub use_avro_logical_types: Option<bool>,
    pub destination_encryption_configuration: Option<EncryptionConfiguration>,
    pub encoding: Option<String>,
    pub quote: Option<String>,
    pub autodetect: Option<bool>,
    pub destination_table: Option<TableReference>,
    pub skip_leading_rows: Option<i32>,
    pub source_uris: Option<Vec<String>>,
    pub null_marker: Option<String>,
    pub time_partitioning: Option<TimePartitioning>,
    pub allow_quoted_newlines: Option<bool>,
    pub projection_fields: Option<Vec<String>>,
    pub clustering: Option<Clustering>,
    pub field_delimiter: Option<String>,
    pub allow_jagged_rows: Option<bool>,
    pub destination_table_properties: Option<DestinationTableProperties>,
    pub source_format: Option<String>,
    pub range_partitioning: Option<RangePartitioning>,
    pub max_bad_records: Option<i32>,
    pub schema_update_options: Option<Vec<String>>,
    pub ignore_unknown_values: Option<bool>,
    pub write_disposition: Option<String>,
    pub hive_partitioning_mode: Option<String>,
    pub hive_partitioning_options: Option<HivePartitioningOptions>,
    pub create_disposition: Option<String>,
    pub schema_inline_format: Option<String>,
    pub schema_inline: Option<String>,
    pub schema: Option<TableSchema>,
}

There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

Fields
use_avro_logical_types: Option<bool>

[Optional] If sourceFormat is set to "AVRO", indicates whether to enable interpreting logical types into their corresponding types (ie. TIMESTAMP), instead of only using their raw types (ie. INTEGER).

destination_encryption_configuration: Option<EncryptionConfiguration>

Custom encryption configuration (e.g., Cloud KMS keys).

encoding: Option<String>

[Optional] The character encoding of the data. The supported values are UTF-8 or ISO-8859-1. The default value is UTF-8. BigQuery decodes the data after the raw, binary data has been split using the values of the quote and fieldDelimiter properties.

quote: Option<String>

[Optional] The value that is used to quote data sections in a CSV file. BigQuery converts the string to ISO-8859-1 encoding, and then uses the first byte of the encoded string to split the data in its raw, binary state. The default value is a double-quote ('"'). If your data does not contain quoted sections, set the property value to an empty string. If your data contains quoted newline characters, you must also set the allowQuotedNewlines property to true.

autodetect: Option<bool>

[Optional] Indicates if we should automatically infer the options and schema for CSV and JSON sources.

destination_table: Option<TableReference>

[Required] The destination table to load the data into.

skip_leading_rows: Option<i32>

[Optional] The number of rows at the top of a CSV file that BigQuery will skip when loading the data. The default value is 0. This property is useful if you have header rows in the file that should be skipped.

source_uris: Option<Vec<String>>

[Required] The fully-qualified URIs that point to your data in Google Cloud. For Google Cloud Storage URIs: Each URI can contain one '' wildcard character and it must come after the 'bucket' name. Size limits related to load jobs apply to external data sources. For Google Cloud Bigtable URIs: Exactly one URI can be specified and it has be a fully specified and valid HTTPS URL for a Google Cloud Bigtable table. For Google Cloud Datastore backups: Exactly one URI can be specified. Also, the '' wildcard character is not allowed.

null_marker: Option<String>

[Optional] Specifies a string that represents a null value in a CSV file. For example, if you specify "\N", BigQuery interprets "\N" as a null value when loading a CSV file. The default value is the empty string. If you set this property to a custom value, BigQuery throws an error if an empty string is present for all data types except for STRING and BYTE. For STRING and BYTE columns, BigQuery interprets the empty string as an empty value.

time_partitioning: Option<TimePartitioning>

Time-based partitioning specification for the destination table. Only one of timePartitioning and rangePartitioning should be specified.

allow_quoted_newlines: Option<bool>

Indicates if BigQuery should allow quoted data sections that contain newline characters in a CSV file. The default value is false.

projection_fields: Option<Vec<String>>

If sourceFormat is set to "DATASTORE_BACKUP", indicates which entity properties to load into BigQuery from a Cloud Datastore backup. Property names are case sensitive and must be top-level properties. If no properties are specified, BigQuery loads all properties. If any named property isn't found in the Cloud Datastore backup, an invalid error is returned in the job result.

clustering: Option<Clustering>

[Beta] Clustering specification for the destination table. Must be specified with time-based partitioning, data in the table will be first partitioned and subsequently clustered.

field_delimiter: Option<String>

[Optional] The separator for fields in a CSV file. The separator can be any ISO-8859-1 single-byte character. To use a character in the range 128-255, you must encode the character as UTF8. BigQuery converts the string to ISO-8859-1 encoding, and then uses the first byte of the encoded string to split the data in its raw, binary state. BigQuery also supports the escape sequence "\t" to specify a tab separator. The default value is a comma (',').

allow_jagged_rows: Option<bool>

[Optional] Accept rows that are missing trailing optional columns. The missing values are treated as nulls. If false, records with missing trailing columns are treated as bad records, and if there are too many bad records, an invalid error is returned in the job result. The default value is false. Only applicable to CSV, ignored for other formats.

destination_table_properties: Option<DestinationTableProperties>

[Beta] [Optional] Properties with which to create the destination table if it is new.

source_format: Option<String>

[Optional] The format of the data files. For CSV files, specify "CSV". For datastore backups, specify "DATASTORE_BACKUP". For newline-delimited JSON, specify "NEWLINE_DELIMITED_JSON". For Avro, specify "AVRO". For parquet, specify "PARQUET". For orc, specify "ORC". The default value is CSV.

range_partitioning: Option<RangePartitioning>

[TrustedTester] Range partitioning specification for this table. Only one of timePartitioning and rangePartitioning should be specified.

max_bad_records: Option<i32>

[Optional] The maximum number of bad records that BigQuery can ignore when running the job. If the number of bad records exceeds this value, an invalid error is returned in the job result. This is only valid for CSV and JSON. The default value is 0, which requires that all records are valid.

schema_update_options: Option<Vec<String>>

Allows the schema of the destination table to be updated as a side effect of the load job if a schema is autodetected or supplied in the job configuration. Schema update options are supported in two cases: when writeDisposition is WRITE_APPEND; when writeDisposition is WRITE_TRUNCATE and the destination table is a partition of a table, specified by partition decorators. For normal tables, WRITE_TRUNCATE will always overwrite the schema. One or more of the following values are specified: ALLOW_FIELD_ADDITION: allow adding a nullable field to the schema. ALLOW_FIELD_RELAXATION: allow relaxing a required field in the original schema to nullable.

ignore_unknown_values: Option<bool>

[Optional] Indicates if BigQuery should allow extra values that are not represented in the table schema. If true, the extra values are ignored. If false, records with extra columns are treated as bad records, and if there are too many bad records, an invalid error is returned in the job result. The default value is false. The sourceFormat property determines what BigQuery treats as an extra value: CSV: Trailing columns JSON: Named values that don't match any column names

write_disposition: Option<String>

[Optional] Specifies the action that occurs if the destination table already exists. The following values are supported: WRITE_TRUNCATE: If the table already exists, BigQuery overwrites the table data. WRITE_APPEND: If the table already exists, BigQuery appends the data to the table. WRITE_EMPTY: If the table already exists and contains data, a 'duplicate' error is returned in the job result. The default value is WRITE_APPEND. Each action is atomic and only occurs if BigQuery is able to complete the job successfully. Creation, truncation and append actions occur as one atomic update upon job completion.

hive_partitioning_mode: Option<String>

[Optional, Trusted Tester] If hive partitioning is enabled, which mode to use. Two modes are supported: - AUTO: automatically infer partition key name(s) and type(s). - STRINGS: automatic infer partition key name(s). All types are strings. Not all storage formats support hive partitioning -- requesting hive partitioning on an unsupported format will lead to an error.

hive_partitioning_options: Option<HivePartitioningOptions>

[Optional, Trusted Tester] Options to configure hive partitioning support.

create_disposition: Option<String>

[Optional] Specifies whether the job is allowed to create new tables. The following values are supported: CREATE_IF_NEEDED: If the table does not exist, BigQuery creates the table. CREATE_NEVER: The table must already exist. If it does not, a 'notFound' error is returned in the job result. The default value is CREATE_IF_NEEDED. Creation, truncation and append actions occur as one atomic update upon job completion.

schema_inline_format: Option<String>

[Deprecated] The format of the schemaInline property.

schema_inline: Option<String>

[Deprecated] The inline schema. For CSV schemas, specify as "Field1:Type1[,Field2:Type2]*". For example, "foo:STRING, bar:INTEGER, baz:FLOAT".

schema: Option<TableSchema>

[Optional] The schema for the destination table. The schema can be omitted if the destination table already exists, or if you're loading data from Google Cloud Datastore.