Struct google_bigquery2::JobConfigurationQuery

pub struct JobConfigurationQuery {
    pub flatten_results: Option<bool>,
    pub use_query_cache: Option<bool>,
    pub destination_table: Option<TableReference>,
    pub query_parameters: Option<Vec<QueryParameter>>,
    pub query: Option<String>,
    pub preserve_nulls: Option<bool>,
    pub maximum_bytes_billed: Option<String>,
    pub maximum_billing_tier: Option<i32>,
    pub parameter_mode: Option<String>,
    pub default_dataset: Option<DatasetReference>,
    pub use_legacy_sql: Option<bool>,
    pub schema_update_options: Option<Vec<String>>,
    pub priority: Option<String>,
    pub write_disposition: Option<String>,
    pub allow_large_results: Option<bool>,
    pub create_disposition: Option<String>,
    pub user_defined_function_resources: Option<Vec<UserDefinedFunctionResource>>,
    pub table_definitions: Option<HashMap<String, ExternalDataConfiguration>>,
}

There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

Fields
flatten_results: Option<bool>

[Optional] Flattens all nested and repeated fields in the query results. The default value is true. allowLargeResults must be true if this is set to false.

use_query_cache: Option<bool>

[Optional] Whether to look for the result in the query cache. The query cache is a best-effort cache that will be flushed whenever tables in the query are modified. Moreover, the query cache is only available when a query does not have a destination table specified. The default value is true.

destination_table: Option<TableReference>

[Optional] Describes the table where the query results should be stored. If not present, a new table will be created to store the results.

query_parameters: Option<Vec<QueryParameter>>

[Experimental] Query parameters for Standard SQL queries.

query: Option<String>

[Required] BigQuery SQL query to execute.

preserve_nulls: Option<bool>

[Deprecated] This property is deprecated.

maximum_bytes_billed: Option<String>

[Optional] Limits the bytes billed for this job. Queries that will have bytes billed beyond this limit will fail (without incurring a charge). If unspecified, this will be set to your project default.

maximum_billing_tier: Option<i32>

[Optional] Limits the billing tier for this job. Queries that have resource usage beyond this tier will fail (without incurring a charge). If unspecified, this will be set to your project default.

parameter_mode: Option<String>

[Experimental] Standard SQL only. Whether to use positional (?) or named (@myparam) query parameters in this query.

default_dataset: Option<DatasetReference>

[Optional] Specifies the default dataset to use for unqualified table names in the query.

use_legacy_sql: Option<bool>

[Experimental] Specifies whether to use BigQuery's legacy SQL dialect for this query. The default value is true. If set to false, the query will use BigQuery's standard SQL: https://cloud.google.com/bigquery/sql-reference/ When useLegacySql is set to false, the values of allowLargeResults and flattenResults are ignored; query will be run as if allowLargeResults is true and flattenResults is false.

schema_update_options: Option<Vec<String>>

[Experimental] Allows the schema of the desitination table to be updated as a side effect of the query job. Schema update options are supported in two cases: when writeDisposition is WRITE_APPEND; when writeDisposition is WRITE_TRUNCATE and the destination table is a partition of a table, specified by partition decorators. For normal tables, WRITE_TRUNCATE will always overwrite the schema. One or more of the following values are specified: ALLOW_FIELD_ADDITION: allow adding a nullable field to the schema. ALLOW_FIELD_RELAXATION: allow relaxing a required field in the original schema to nullable.

priority: Option<String>

[Optional] Specifies a priority for the query. Possible values include INTERACTIVE and BATCH. The default value is INTERACTIVE.

write_disposition: Option<String>

[Optional] Specifies the action that occurs if the destination table already exists. The following values are supported: WRITE_TRUNCATE: If the table already exists, BigQuery overwrites the table data. WRITE_APPEND: If the table already exists, BigQuery appends the data to the table. WRITE_EMPTY: If the table already exists and contains data, a 'duplicate' error is returned in the job result. The default value is WRITE_EMPTY. Each action is atomic and only occurs if BigQuery is able to complete the job successfully. Creation, truncation and append actions occur as one atomic update upon job completion.

allow_large_results: Option<bool>

If true, allows the query to produce arbitrarily large result tables at a slight cost in performance. Requires destinationTable to be set.

create_disposition: Option<String>

[Optional] Specifies whether the job is allowed to create new tables. The following values are supported: CREATE_IF_NEEDED: If the table does not exist, BigQuery creates the table. CREATE_NEVER: The table must already exist. If it does not, a 'notFound' error is returned in the job result. The default value is CREATE_IF_NEEDED. Creation, truncation and append actions occur as one atomic update upon job completion.

user_defined_function_resources: Option<Vec<UserDefinedFunctionResource>>

[Experimental] Describes user-defined function resources used in the query.

table_definitions: Option<HashMap<String, ExternalDataConfiguration>>

[Optional] If querying an external data source outside of BigQuery, describes the data format, location and other properties of the data source. By defining these properties, the data source can then be queried as if it were a standard BigQuery table.