Struct google_bigquery2::JobConfigurationTableCopy

pub struct JobConfigurationTableCopy {
    pub create_disposition: Option<String>,
    pub source_tables: Option<Vec<TableReference>>,
    pub write_disposition: Option<String>,
    pub destination_table: Option<TableReference>,
    pub source_table: Option<TableReference>,
}

There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

Fields
create_disposition: Option<String>

[Optional] Specifies whether the job is allowed to create new tables. The following values are supported: CREATE_IF_NEEDED: If the table does not exist, BigQuery creates the table. CREATE_NEVER: The table must already exist. If it does not, a 'notFound' error is returned in the job result. The default value is CREATE_IF_NEEDED. Creation, truncation and append actions occur as one atomic update upon job completion.

source_tables: Option<Vec<TableReference>>

[Pick one] Source tables to copy.

write_disposition: Option<String>

[Optional] Specifies the action that occurs if the destination table already exists. The following values are supported: WRITE_TRUNCATE: If the table already exists, BigQuery overwrites the table data. WRITE_APPEND: If the table already exists, BigQuery appends the data to the table. WRITE_EMPTY: If the table already exists and contains data, a 'duplicate' error is returned in the job result. The default value is WRITE_EMPTY. Each action is atomic and only occurs if BigQuery is able to complete the job successfully. Creation, truncation and append actions occur as one atomic update upon job completion.

destination_table: Option<TableReference>

[Required] The destination table

source_table: Option<TableReference>

[Pick one] Source table to copy.