Struct google_bigquery2::JobMethods

pub struct JobMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on job resources. It is not used directly, but through the Bigquery hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_bigquery2 as bigquery2;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use bigquery2::Bigquery;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Bigquery::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `get(...)`, `get_query_results(...)`, `insert(...)`, `list(...)` and `query(...)`
// to build up your call.
let rb = hub.jobs();

Methods

impl<'a, C, A> JobMethods<'a, C, A>

fn query(&self, request: QueryRequest, project_id: &str) -> JobQueryCall<'a, C, A>

Create a builder to help you perform the following task:

Runs a BigQuery SQL query synchronously and returns query results if the query completes within a specified timeout.

Arguments

request - No description provided.

projectId - Project ID of the project billed for the query

fn get_query_results(&self, project_id: &str, job_id: &str) -> JobGetQueryResultCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves the results of a query job.

Arguments

projectId - Project ID of the query job

jobId - Job ID of the query job

fn list(&self, project_id: &str) -> JobListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists all jobs that you started in the specified project. The job list returns in reverse chronological order of when the jobs were created, starting with the most recent job created. Requires the Can View project role, or the Is Owner project role if you set the allUsers property.

Arguments

projectId - Project ID of the jobs to list

fn get(&self, project_id: &str, job_id: &str) -> JobGetCall<'a, C, A>

Create a builder to help you perform the following task:

Returns information about a specific job. Job information is available for a six month period after creation. Requires that you're the person who ran the job, or have the Is Owner project role.

Arguments

projectId - Project ID of the requested job

jobId - Job ID of the requested job

fn insert(&self, request: Job, project_id: &str) -> JobInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Starts a new asynchronous job. Requires the Can View project role.

Arguments

request - No description provided.

projectId - Project ID of the project that will be billed for the job