Struct google_bigquery2::JobStatistics

pub struct JobStatistics {
    pub load: Option<JobStatistics3>,
    pub parent_job_id: Option<String>,
    pub num_child_jobs: Option<String>,
    pub total_slot_ms: Option<String>,
    pub completion_ratio: Option<f64>,
    pub extract: Option<JobStatistics4>,
    pub reservation_usage: Option<Vec<JobStatisticsReservationUsage>>,
    pub creation_time: Option<String>,
    pub total_bytes_processed: Option<String>,
    pub quota_deferments: Option<Vec<String>>,
    pub start_time: Option<String>,
    pub query: Option<JobStatistics2>,
    pub end_time: Option<String>,
    pub reservation_id: Option<String>,
}

There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

Fields
load: Option<JobStatistics3>

[Output-only] Statistics for a load job.

parent_job_id: Option<String>

[Output-only] If this is a child job, the id of the parent.

num_child_jobs: Option<String>

[Output-only] Number of child jobs executed.

total_slot_ms: Option<String>

[Output-only] Slot-milliseconds for the job.

completion_ratio: Option<f64>

[TrustedTester] [Output-only] Job progress (0.0 -> 1.0) for LOAD and EXTRACT jobs.

extract: Option<JobStatistics4>

[Output-only] Statistics for an extract job.

reservation_usage: Option<Vec<JobStatisticsReservationUsage>>

[Output-only] Job resource usage breakdown by reservation.

creation_time: Option<String>

[Output-only] Creation time of this job, in milliseconds since the epoch. This field will be present on all jobs.

total_bytes_processed: Option<String>

[Output-only] [Deprecated] Use the bytes processed in the query statistics instead.

quota_deferments: Option<Vec<String>>

[Output-only] Quotas which delayed this job's start time.

start_time: Option<String>

[Output-only] Start time of this job, in milliseconds since the epoch. This field will be present when the job transitions from the PENDING state to either RUNNING or DONE.

query: Option<JobStatistics2>

[Output-only] Statistics for a query job.

end_time: Option<String>

[Output-only] End time of this job, in milliseconds since the epoch. This field will be present whenever a job is in the DONE state.

reservation_id: Option<String>

[Output-only] Name of the primary reservation assigned to this job. Note that this could be different than reservations reported in the reservation usage field if parent reservations were used to execute this job.