Struct google_bigquery2::JobStatistics2

pub struct JobStatistics2 {
    pub query_plan: Option<Vec<ExplainQueryStage>>,
    pub total_bytes_billed: Option<String>,
    pub total_bytes_processed: Option<String>,
    pub cache_hit: Option<bool>,
    pub undeclared_query_parameters: Option<Vec<QueryParameter>>,
    pub referenced_tables: Option<Vec<TableReference>>,
    pub billing_tier: Option<i32>,
    pub num_dml_affected_rows: Option<String>,
    pub schema: Option<TableSchema>,
}

There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

Fields
query_plan: Option<Vec<ExplainQueryStage>>

[Output-only, Experimental] Describes execution plan for the query.

total_bytes_billed: Option<String>

[Output-only] Total bytes billed for the job.

total_bytes_processed: Option<String>

[Output-only] Total bytes processed for the job.

cache_hit: Option<bool>

[Output-only] Whether the query result was fetched from the query cache.

undeclared_query_parameters: Option<Vec<QueryParameter>>

[Output-only, Experimental] Standard SQL only: list of undeclared query parameters detected during a dry run validation.

referenced_tables: Option<Vec<TableReference>>

[Output-only, Experimental] Referenced tables for the job. Queries that reference more than 50 tables will not have a complete list.

billing_tier: Option<i32>

[Output-only] Billing tier for the job.

num_dml_affected_rows: Option<String>

[Output-only, Experimental] The number of rows affected by a DML statement. Present only for DML statements INSERT, UPDATE or DELETE.

schema: Option<TableSchema>

[Output-only, Experimental] The schema of the results. Present only for successful dry run of non-legacy SQL queries.