Struct google_bigquery2::QueryResponse

pub struct QueryResponse {
    pub kind: Option<String>,
    pub errors: Option<Vec<ErrorProto>>,
    pub job_reference: Option<JobReference>,
    pub cache_hit: Option<bool>,
    pub job_complete: Option<bool>,
    pub total_rows: Option<String>,
    pub total_bytes_processed: Option<String>,
    pub page_token: Option<String>,
    pub rows: Option<Vec<TableRow>>,
    pub schema: Option<TableSchema>,
}

There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

query jobs (response)

Fields

kind

The resource type.

errors

[Output-only] All errors and warnings encountered during the running of the job. Errors here do not necessarily mean that the job has completed or was unsuccessful.

job_reference

Reference to the Job that was created to run the query. This field will be present even if the original request timed out, in which case GetQueryResults can be used to read the results once the query has completed. Since this API only returns the first page of results, subsequent pages can be fetched via the same mechanism (GetQueryResults).

cache_hit

Whether the query result was fetched from the query cache.

job_complete

Whether the query has completed or not. If rows or totalRows are present, this will always be true. If this is false, totalRows will not be available.

total_rows

The total number of rows in the complete query result set, which can be more than the number of rows in this single page of results.

total_bytes_processed

The total number of bytes processed for this query. If this query was a dry run, this is the number of bytes that would be processed if the query were run.

page_token

A token used for paging results.

rows

An object with as many results as can be contained within the maximum permitted reply size. To get any additional rows, you can call GetQueryResults and specify the jobReference returned above.

schema

The schema of the results. Present only when the query completes successfully.