Struct google_bigquery2::Table

pub struct Table {
    pub description: Option<String>,
    pub creation_time: Option<String>,
    pub num_bytes: Option<String>,
    pub time_partitioning: Option<TimePartitioning>,
    pub last_modified_time: Option<String>,
    pub id: Option<String>,
    pub kind: Option<String>,
    pub streaming_buffer: Option<Streamingbuffer>,
    pub external_data_configuration: Option<ExternalDataConfiguration>,
    pub location: Option<String>,
    pub table_reference: Option<TableReference>,
    pub num_rows: Option<String>,
    pub etag: Option<String>,
    pub num_long_term_bytes: Option<String>,
    pub friendly_name: Option<String>,
    pub expiration_time: Option<String>,
    pub view: Option<ViewDefinition>,
    pub type_: Option<String>,
    pub self_link: Option<String>,
    pub schema: Option<TableSchema>,
}

There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

update tables (request|response)

insert tables (request|response)

list tables (none)

delete tables (none)

get tables (response)

patch tables (request|response)

Fields
description: Option<String>

[Optional] A user-friendly description of this table.

creation_time: Option<String>

[Output-only] The time when this table was created, in milliseconds since the epoch.

num_bytes: Option<String>

[Output-only] The size of this table in bytes, excluding any data in the streaming buffer.

time_partitioning: Option<TimePartitioning>

[Experimental] If specified, configures time-based partitioning for this table.

last_modified_time: Option<String>

[Output-only] The time when this table was last modified, in milliseconds since the epoch.

id: Option<String>

[Output-only] An opaque ID uniquely identifying the table.

kind: Option<String>

[Output-only] The type of the resource.

streaming_buffer: Option<Streamingbuffer>

[Output-only] Contains information regarding this table's streaming buffer, if one is present. This field will be absent if the table is not being streamed to or if there is no data in the streaming buffer.

external_data_configuration: Option<ExternalDataConfiguration>

[Optional] Describes the data format, location, and other properties of a table stored outside of BigQuery. By defining these properties, the data source can then be queried as if it were a standard BigQuery table.

location: Option<String>

[Output-only] The geographic location where the table resides. This value is inherited from the dataset.

table_reference: Option<TableReference>

[Required] Reference describing the ID of this table.

num_rows: Option<String>

[Output-only] The number of rows of data in this table, excluding any data in the streaming buffer.

etag: Option<String>

[Output-only] A hash of this resource.

num_long_term_bytes: Option<String>

[Output-only] The number of bytes in the table that are considered "long-term storage".

friendly_name: Option<String>

[Optional] A descriptive name for this table.

expiration_time: Option<String>

[Optional] The time when this table expires, in milliseconds since the epoch. If not present, the table will persist indefinitely. Expired tables will be deleted and their storage reclaimed.

view: Option<ViewDefinition>

[Optional] The view definition.

type_: Option<String>

[Output-only] Describes the table type. The following values are supported: TABLE: A normal BigQuery table. VIEW: A virtual table defined by a SQL query. EXTERNAL: A table that references data stored in an external storage system, such as Google Cloud Storage. The default value is TABLE.

self_link: Option<String>

[Output-only] A URL that can be used to access this resource again.

schema: Option<TableSchema>

[Optional] Describes the schema of this table.