Struct google_bigquery2::TabledataMethods

pub struct TabledataMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on tabledata resources. It is not used directly, but through the Bigquery hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_bigquery2 as bigquery2;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use bigquery2::Bigquery;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Bigquery::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `insert_all(...)` and `list(...)`
// to build up your call.
let rb = hub.tabledata();

Methods

impl<'a, C, A> TabledataMethods<'a, C, A>

fn insert_all(&self, request: TableDataInsertAllRequest, project_id: &str, dataset_id: &str, table_id: &str) -> TabledataInsertAllCall<'a, C, A>

Create a builder to help you perform the following task:

Streams data into BigQuery one record at a time without needing to run a load job. Requires the WRITER dataset role.

Arguments

request - No description provided.

projectId - Project ID of the destination table.

datasetId - Dataset ID of the destination table.

tableId - Table ID of the destination table.

fn list(&self, project_id: &str, dataset_id: &str, table_id: &str) -> TabledataListCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves table data from a specified set of rows. Requires the READER dataset role.

Arguments

projectId - Project ID of the table to read

datasetId - Dataset ID of the table to read

tableId - Table ID of the table to read