Struct google_bigquery2::Dataset

pub struct Dataset {
    pub kind: Option<String>,
    pub description: Option<String>,
    pub dataset_reference: Option<DatasetReference>,
    pub creation_time: Option<String>,
    pub access: Option<Vec<DatasetAccess>>,
    pub default_table_expiration_ms: Option<String>,
    pub etag: Option<String>,
    pub location: Option<String>,
    pub friendly_name: Option<String>,
    pub last_modified_time: Option<String>,
    pub id: Option<String>,
    pub self_link: Option<String>,
}

There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get datasets (response)

list datasets (none)

patch datasets (request|response)

update datasets (request|response)

delete datasets (none)

insert datasets (request|response)

Fields

kind

[Output-only] The resource type.

description

[Optional] A user-friendly description of the dataset.

dataset_reference

[Required] A reference that identifies the dataset.

creation_time

[Output-only] The time when this dataset was created, in milliseconds since the epoch.

access

[Optional] An array of objects that define dataset access for one or more entities. You can set this property when inserting or updating a dataset in order to control who is allowed to access the data. If unspecified at dataset creation time, BigQuery adds default dataset access for the following entities: access.specialGroup: projectReaders; access.role: READER; access.specialGroup: projectWriters; access.role: WRITER; access.specialGroup: projectOwners; access.role: OWNER; access.userByEmail: [dataset creator email]; access.role: OWNER;

default_table_expiration_ms

[Optional] The default lifetime of all tables in the dataset, in milliseconds. The minimum value is 3600000 milliseconds (one hour). Once this property is set, all newly-created tables in the dataset will have an expirationTime property set to the creation time plus the value in this property, and changing the value will only affect new tables, not existing ones. When the expirationTime for a given table is reached, that table will be deleted automatically. If a table's expirationTime is modified or removed before the table expires, or if you provide an explicit expirationTime when creating a table, that value takes precedence over the default expiration time indicated by this property.

etag

[Output-only] A hash of the resource.

location

[Experimental] The geographic location where the dataset should reside. Possible values include EU and US. The default value is US.

friendly_name

[Optional] A descriptive name for the dataset.

last_modified_time

[Output-only] The date when this dataset or any of its tables was last modified, in milliseconds since the epoch.

id

[Output-only] The fully-qualified unique name of the dataset in the format projectId:datasetId. The dataset name without the project name is given in the datasetId field. When creating a new dataset, leave this field blank, and instead specify the datasetId field.

self_link

[Output-only] A URL that can be used to access the resource again. You can use this URL in Get or Update requests to the resource.