Struct google_bigquery2::ExplainQueryStage

pub struct ExplainQueryStage {
    pub compute_ratio_max: Option<f64>,
    pub name: Option<String>,
    pub records_written: Option<String>,
    pub wait_ratio_max: Option<f64>,
    pub read_ratio_avg: Option<f64>,
    pub read_ratio_max: Option<f64>,
    pub steps: Option<Vec<ExplainQueryStep>>,
    pub records_read: Option<String>,
    pub compute_ratio_avg: Option<f64>,
    pub write_ratio_max: Option<f64>,
    pub wait_ratio_avg: Option<f64>,
    pub id: Option<String>,
    pub write_ratio_avg: Option<f64>,
}

There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

Fields

compute_ratio_max

Relative amount of time the slowest shard spent on CPU-bound tasks.

name

Human-readable name for stage.

records_written

Number of records written by the stage.

wait_ratio_max

Relative amount of time the slowest shard spent waiting to be scheduled.

read_ratio_avg

Relative amount of time the average shard spent reading input.

read_ratio_max

Relative amount of time the slowest shard spent reading input.

steps

List of operations within the stage in dependency order (approximately chronological).

records_read

Number of records read into the stage.

compute_ratio_avg

Relative amount of time the average shard spent on CPU-bound tasks.

write_ratio_max

Relative amount of time the slowest shard spent on writing output.

wait_ratio_avg

Relative amount of time the average shard spent waiting to be scheduled.

id

Unique ID for stage within plan.

write_ratio_avg

Relative amount of time the average shard spent on writing output.