Struct google_bigquery2::ExternalDataConfiguration

pub struct ExternalDataConfiguration {
    pub ignore_unknown_values: Option<bool>,
    pub source_uris: Option<Vec<String>>,
    pub compression: Option<String>,
    pub csv_options: Option<CsvOptions>,
    pub source_format: Option<String>,
    pub max_bad_records: Option<i32>,
    pub schema: Option<TableSchema>,
}

There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

Fields

ignore_unknown_values

[Optional] Indicates if BigQuery should allow extra values that are not represented in the table schema. If true, the extra values are ignored. If false, records with extra columns are treated as bad records, and if there are too many bad records, an invalid error is returned in the job result. The default value is false. The sourceFormat property determines what BigQuery treats as an extra value: CSV: Trailing columns JSON: Named values that don't match any column names Google Cloud Datastore backups: This setting is ignored.

source_uris

[Required] The fully-qualified URIs that point to your data in Google Cloud Storage. Each URI can contain one '*' wildcard character and it must come after the 'bucket' name. Size limits related to load jobs apply to external data sources, plus an additional limit of 10 GB maximum size across all URIs. For Google Cloud Datastore backups, exactly one URI can be specified, and it must end with '.backup_info'. Also, the '*' wildcard character is not allowed.

compression

[Optional] The compression type of the data source. Possible values include GZIP and NONE. The default value is NONE. This setting is ignored for Google Cloud Datastore backups.

csv_options

Additional properties to set if sourceFormat is set to CSV.

source_format

[Required] The data format. For CSV files, specify "CSV". For newline-delimited JSON, specify "NEWLINE_DELIMITED_JSON". For Google Cloud Datastore backups, specify "DATASTORE_BACKUP".

max_bad_records

[Optional] The maximum number of bad records that BigQuery can ignore when reading data. If the number of bad records exceeds this value, an invalid error is returned in the job result. The default value is 0, which requires that all records are valid. This setting is ignored for Google Cloud Datastore backups.

schema

[Optional] The schema for the data. Schema is required for CSV and JSON formats. Schema is disallowed for Google Cloud Datastore backups.