Struct google_bigquery2::GetQueryResultsResponse

pub struct GetQueryResultsResponse {
    pub kind: Option<String>,
    pub errors: Option<Vec<ErrorProto>>,
    pub job_reference: Option<JobReference>,
    pub cache_hit: Option<bool>,
    pub job_complete: Option<bool>,
    pub total_rows: Option<String>,
    pub total_bytes_processed: Option<String>,
    pub page_token: Option<String>,
    pub etag: Option<String>,
    pub rows: Option<Vec<TableRow>>,
    pub num_dml_affected_rows: Option<String>,
    pub schema: Option<TableSchema>,
}

There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get query results jobs (response)

Fields
kind: Option<String>

The resource type of the response.

errors: Option<Vec<ErrorProto>>

[Output-only] All errors and warnings encountered during the running of the job. Errors here do not necessarily mean that the job has completed or was unsuccessful.

job_reference: Option<JobReference>

Reference to the BigQuery Job that was created to run the query. This field will be present even if the original request timed out, in which case GetQueryResults can be used to read the results once the query has completed. Since this API only returns the first page of results, subsequent pages can be fetched via the same mechanism (GetQueryResults).

cache_hit: Option<bool>

Whether the query result was fetched from the query cache.

job_complete: Option<bool>

Whether the query has completed or not. If rows or totalRows are present, this will always be true. If this is false, totalRows will not be available.

total_rows: Option<String>

The total number of rows in the complete query result set, which can be more than the number of rows in this single page of results. Present only when the query completes successfully.

total_bytes_processed: Option<String>

The total number of bytes processed for this query.

page_token: Option<String>

A token used for paging results.

etag: Option<String>

A hash of this response.

rows: Option<Vec<TableRow>>

An object with as many results as can be contained within the maximum permitted reply size. To get any additional rows, you can call GetQueryResults and specify the jobReference returned above. Present only when the query completes successfully.

num_dml_affected_rows: Option<String>

[Output-only, Experimental] The number of rows affected by a DML statement. Present only for DML statements INSERT, UPDATE or DELETE.

schema: Option<TableSchema>

The schema of the results. Present only when the query completes successfully.