Struct google_bigquery2::JobConfigurationQuery

pub struct JobConfigurationQuery {
    pub flatten_results: Option<bool>,
    pub use_query_cache: Option<bool>,
    pub default_dataset: Option<DatasetReference>,
    pub destination_table: Option<TableReference>,
    pub priority: Option<String>,
    pub write_disposition: Option<String>,
    pub allow_large_results: Option<bool>,
    pub create_disposition: Option<String>,
    pub query: Option<String>,
    pub preserve_nulls: Option<bool>,
    pub table_definitions: Option<HashMap<String, ExternalDataConfiguration>>,
}

There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

Fields

flatten_results

[Optional] Flattens all nested and repeated fields in the query results. The default value is true. allowLargeResults must be true if this is set to false.

use_query_cache

[Optional] Whether to look for the result in the query cache. The query cache is a best-effort cache that will be flushed whenever tables in the query are modified. Moreover, the query cache is only available when a query does not have a destination table specified.

default_dataset

[Optional] Specifies the default dataset to use for unqualified table names in the query.

destination_table

[Optional] Describes the table where the query results should be stored. If not present, a new table will be created to store the results.

priority

[Optional] Specifies a priority for the query. Possible values include INTERACTIVE and BATCH. The default value is INTERACTIVE.

write_disposition

[Optional] Specifies the action that occurs if the destination table already exists. The following values are supported: WRITE_TRUNCATE: If the table already exists, BigQuery overwrites the table data. WRITE_APPEND: If the table already exists, BigQuery appends the data to the table. WRITE_EMPTY: If the table already exists and contains data, a 'duplicate' error is returned in the job result. The default value is WRITE_EMPTY. Each action is atomic and only occurs if BigQuery is able to complete the job successfully. Creation, truncation and append actions occur as one atomic update upon job completion.

allow_large_results

If true, allows the query to produce arbitrarily large result tables at a slight cost in performance. Requires destinationTable to be set.

create_disposition

[Optional] Specifies whether the job is allowed to create new tables. The following values are supported: CREATE_IF_NEEDED: If the table does not exist, BigQuery creates the table. CREATE_NEVER: The table must already exist. If it does not, a 'notFound' error is returned in the job result. The default value is CREATE_IF_NEEDED. Creation, truncation and append actions occur as one atomic update upon job completion.

query

[Required] BigQuery SQL query to execute.

preserve_nulls

[Deprecated] This property is deprecated.

table_definitions

[Experimental] If querying an external data source outside of BigQuery, describes the data format, location and other properties of the data source. By defining these properties, the data source can then be queried as if it were a standard BigQuery table.