Struct google_bigquery2::QueryRequest

pub struct QueryRequest {
    pub timeout_ms: Option<u32>,
    pub kind: Option<String>,
    pub dry_run: Option<bool>,
    pub parameter_mode: Option<String>,
    pub use_query_cache: Option<bool>,
    pub default_dataset: Option<DatasetReference>,
    pub use_legacy_sql: Option<bool>,
    pub max_results: Option<u32>,
    pub query_parameters: Option<Vec<QueryParameter>>,
    pub query: Option<String>,
    pub preserve_nulls: Option<bool>,
}

There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

query jobs (request)

Fields
timeout_ms: Option<u32>

[Optional] How long to wait for the query to complete, in milliseconds, before the request times out and returns. Note that this is only a timeout for the request, not the query. If the query takes longer to run than the timeout value, the call returns without any results and with the 'jobComplete' flag set to false. You can call GetQueryResults() to wait for the query to complete and read the results. The default value is 10000 milliseconds (10 seconds).

kind: Option<String>

The resource type of the request.

dry_run: Option<bool>

[Optional] If set to true, BigQuery doesn't run the job. Instead, if the query is valid, BigQuery returns statistics about the job such as how many bytes would be processed. If the query is invalid, an error returns. The default value is false.

parameter_mode: Option<String>

[Experimental] Standard SQL only. Whether to use positional (?) or named (@myparam) query parameters in this query.

use_query_cache: Option<bool>

[Optional] Whether to look for the result in the query cache. The query cache is a best-effort cache that will be flushed whenever tables in the query are modified. The default value is true.

default_dataset: Option<DatasetReference>

[Optional] Specifies the default datasetId and projectId to assume for any unqualified table names in the query. If not set, all table names in the query string must be qualified in the format 'datasetId.tableId'.

use_legacy_sql: Option<bool>

[Experimental] Specifies whether to use BigQuery's legacy SQL dialect for this query. The default value is true. If set to false, the query will use BigQuery's standard SQL: https://cloud.google.com/bigquery/sql-reference/ When useLegacySql is set to false, the values of allowLargeResults and flattenResults are ignored; query will be run as if allowLargeResults is true and flattenResults is false.

max_results: Option<u32>

[Optional] The maximum number of rows of data to return per page of results. Setting this flag to a small value such as 1000 and then paging through results might improve reliability when the query result set is large. In addition to this limit, responses are also limited to 10 MB. By default, there is no maximum row count, and only the byte limit applies.

query_parameters: Option<Vec<QueryParameter>>

[Experimental] Query parameters for Standard SQL queries.

query: Option<String>

[Required] A query string, following the BigQuery query syntax, of the query to execute. Example: "SELECT count(f1) FROM [myProjectId:myDatasetId.myTableId]".

preserve_nulls: Option<bool>

[Deprecated] This property is deprecated.